refactor workers execution composition seams
