+++ docs/internal/baselines/backend-package-file-count.json
@@ -129,7 +129,7 @@
     },
     {
       "packagePath": "pkg/services/factory_sessions/internal/runtimeopening",
-      "fileCount": 27,
+      "fileCount": 26,
       "owner": "factory-session-maintainers",
       "removalReason": "Partition runtime opening composition, model binding, and runtime construction by responsibility."
     },
