+++ pkg/services/factory_runtime/internal/build.go
@@ -420,6 +420,7 @@ func assembleRuntimeBundle(
 		workService,
 		workRequestIDs,
 		newID,
+		runtimeDirs,
 		decisionEnvelopes,
 	)
 	if err != nil {
